i#4549 Github Actions: Trigger on all pushes, not on edits
Changes the Github Actions trigger to remove the pull request "edited"
action, which is triggered on any top-level comment, title changes,
etc. and is thus more than we want.  Instead we trigger on a push to
any branch, plus pull request creation and "synchronize" which seems
to cover code updates to the pull request branch even when in a forked
repo.

Issue: #4549