i#4549 Github Actions: Do not trigger on edits (#4580)
Changes the Github Actions trigger to remove the pull request "edited"
action, which is triggered on some top-level comments, title changes,
etc. and is thus more than we want.  The pull request "synchronize"
trigger is enough to cover updates to a pull request branch even
when in a forked repo.

Issue: #4549