i#4588: Handle anomalous alloc bases in Windows query loop
Removes what was thought of as a sanity check for the allocation base
changing in DR's internal query loop, but it turns out there are cases
of anomalous bases for which failing the query has disastrous
consequences.  Just ignoring the anomaly and moving on is the
solution.

Issue: #4588, DynamoRIO/drmemory#2328
Fixes #4588