i#4131: Set up Windows Github Actions (#4586)
Adds three separate jobs on Windows: 32-bit debug with tests, 64-bit debug with tests,
and a job that creates 3 builds without tests: release-32, release-64, and vps-32-debug.
Adds corresponding flags to the script layers to accomplish this build splitting.

We target a VS2017 Win10-1607 image to match what we had on Appveyor.

Modifies the runsuite_wrapper.pl script to *not* fork, since there is
no Cygwin perl available by default.  Instead, we tee to a file.
Updates checks for Cygwin to apply to native Windows perl as well.

Includes attempts to improve two tests:
+ Add error code printing for drx-test
+ Generalize api.symtest golden output to handle extra operators found

Other failing tests were either already on the ignore list or we add them here.
The long list of tests ignored for #4058 is culled to remove tests never observed
to fail on GA.

Removes APPVEYOR_PULL_REQUEST_NUMBER-driven build skips now that we have
a separate build job.  Sets a CI_TRIGGER env var and use it to shrink 32-bit tests in the
same manner as before for Appveyor PR runs.

This is based on the similar work for Dr. Memory:
https://github.com/DynamoRIO/drmemory/pull/2326

Issue: #4131