iX: add platform-independent macros for call ind (#4630)
Adds XINST_CREATE_call_reg macros for ARM, AArch64 and x86. They point to the respective instrs on those platforms for transferring control to address in given register operand.

Adds tests to verify XINST_CREATE_call_reg encoding. But, expected test output for ARM is missing, because we don't have an ARM machine to run it on and verify. Added a TODO for that.