i#4549 GA CI: Fix cronbuild version errors and allow manual
The cronbuild version was missing the major and minor: that was on the
tag only.  We fix that here.

Adds a source checkout to the release job, needed to run "git log"
to compute the version.

Allows manual cronbuilds via a blank version, so we can actually test
them on demand, and create them separately from the scheduled time.

Applies input build numbers to cronbuilds too so we can manually test
them the same day as a scheduled build.

Moves the scheduled cronbuild to 2am UTC for 9pm EST.

Issue: #4549