i#1921 native sig: Account for sigaltstack when placing native frames (#4645)
The native frame adjustments to handle the DR handler being on the
native stack previously failed to handle an application with a
sigaltstack.  We correct that here.

Adds sigaltstack testing to the api.detach_signal test.  With a
smaller stack, this exercises the stack-too-small error message path
from DR.

Issue: #1921