i#4658: Fix fragment_prefix_size on AArch64. (#4659)
fragment_prefix_size returns an incorrect value on AArch64 when -shared_bb_ibt_tables is set. To fix this, we set the default value of -bb_prefixes to true on AArch64, where we do have fragment prefixes.

Also adds a rule to verify that -bb_prefixes is always set on AArch64.