Enable ibl-stress test on AArch64.
This test with some options helps trigger the far fragment linking path on AArch64. Unfortunately, with those options, the test is too big to enable in CI. So for the CI, we use a smaller version of the test, which doesn't help test the far fragment linking path, but is good to enable on AArch64 anyway. The larger version of the test as specified by the conditionally defined constants in ibl-stress.c can be used to invoke/test the far fragment linking path on AArch64.