i#4652: Stop searching for additional prefixes in decode_cti after a c5 VEX prefix is found
When decoding the two-byte c5 VEX prefix, currently the second byte of the prefix is not skipped and is instead treated as its own prefix. If the second byte is, say, c4, that will trigger the three-byte c4 VEX prefix code path which will cause decode_cti to fail spuriously. Since VEX prefixes are always the final prefix, just copy what we do for c4 VEX and EVEX prefixes and stop looking for additional prefixes after we see a c5 VEX prefix.

Issue: #4652