i#4669 xl8 failures: No asserts for client xl8 requests
Applies the existing client_data->is_translating debug flag to avoid a
curiosity and an assert on translation failure stemming from client
use of dr_app_pc_from_cache_pc() on a sampling itimer.

Augments the client.timer test to run with -disable_traces and use a
lot of indirect branches to trigger a curiosity on xl8 failure on
every run when the iterations are increased (a little too high for our
small regression tests though).  With this fix, the curiosity never
occurs.

Issue: #4669