i#4680 aarch xl8: Fix GPR restore bug on cross-app spills
Fixes a translation bug in translate_walk_restore() where it just
skips GPR restores if it's in between application instructions.  Since
on AArchXX we have many of these cases of cross-app-instr spills (due
to restore elision for back-to-back spills), the assumption that there
will be no outstanding spills does not hold.

The solution is to move the condition to only encompass the stack
pointer restore, while the GPR restores should always be done.

This was caught by the forthcoming common.broadfun-stress test, which
will serve as the regression test once it is added after all other
bugs it finds are addressed.

Issue: #4680