i#4665: Fix crash after IBT resize on AArch64
Adds missing restore for next_tag and last_exit linkstub in indirect branch lookups after IBT resize. In this case, the IBT payloads are replaced with the target_delete_entry gencode address to induce a cache exit, and safely nullify the table.

Removes the -shared_ibt_table_bb_init workaround for ibl-stress test on AArch64, which was required to avoid the buggy IBT resize.

Verified locally that the ibl-stress-aarch64-far-link_LONG test also works without the workaround after this fix. Manual local verification was required as this long test is not enabled in the regular suite.

Fixes: #4665