i#4717: Fix hang in api.static_prepop test on aarchxx
Fixes a bug in the api.static_prepop assembly where the link register
is clobbered, leading to an infinite loop on aarchxx.
Enables the test for aarchxx.

Fixes a bug on identifying stopping points on ARM where LSB Thumb
decoration on the two sides of the comparison was not consistent.
Fixes a corresponding bug on going native where the LSB was not set.

Tested manually on ARM.

Issue: #1578, #4717, #4720
Fixes #4717