i#4719 qemu: Add -takeover_timeout_ms
Fixes a problem with the hardcoded small timeout from PR #4725 by
parameterizing the timeout in a new option -takeover_timeout_ms.  It
is set to a high value by default; the plan is to have -xarch_root set
it to a low value for the common QEMU case of running a small test,
while still overridable for large apps.

Renames -ignore_takeover_timeout to -unsafe_ignore_takeover_timeout to
indicate that it can cause problems if actual application threads are
left native.

Issue: #4719