i#4627: Work around QEMU signal number bugs
QEMU prior to 5.0 has a signal number mapping bug where it ends up
failing its app's attempts to invoke sigaction for signals 33 and 64.
We relax DR's assert on sigaction failure when -xarch_root is set for
those numbers, trying to continue for best-effort execution with a
syslog warning.

Tested manually on 4.2.0 and on the forthcoming GA CI QEMU testing in
PR #4747.

Fixes #4627