i#4111 web: Reduce docs cronbuilds to weekly
The daily docs cronbuilds are steadily increasing the size of the
repo, due to timestamps in every single file changing every time.
Given that we rarely make docs changes that we want to make live
immediately, we reduce the cronbuild to coincide with the package
build, and rely on manually-triggered docs builds for anything else.

Issue: #4111