i#4670 signal-interrupted special ibl xfer frag: Unlink if signal pending
Adds check for pending signals in the special ibl xfer trampoline. If there are pending signals, we transfer control to the unlinked ibl entry point instead.

This is required to ensure that we exit the fcache and bound delivery time for signals received while executing fragments that use the special ibl xfer mechanism.

Fixes: #4670