i#4769: Reset global state of ext libs for reattaches with static-DR.
Resets part of the global state of several ext libs when detaching is in progress.
It also adds a DR API routine to query whether a detach is in progress.
For testing, adds a regression test that re-attaches with static-DR 100x.
This test is based on burst_malloc and it takes 270sec (21% faster than burst_malloc with 100 reattaches and # prefix on the ops).

Fixes: #4769