iX: remove two cases of ASSERT_NOT_TESTED() for AARCH64
PR #3003 added handling of signal interrupting the ibl hit path's jump, but the introduced
client.cleancallsig test could not test the ibl jump on AArchXX leading to the insertion of a
ASSERT_NOT_TESTED() for AARCHXX. Testing on proprietary application with plain DR (debug build)
on AArch64 hit this case without problems. Thus, we can remove the ASSERT_NOT_TESTED() for AARCH64.

The code inserted in PR #4797 was triggered and tested for AARCH64 on a proprietary app so removing
the ASSERT_NOT_TESTED() for AARCH64 from there as well.