Remove two (now tested) cases of ASSERT_NOT_TESTED for AARCH64 (#4839)
PR #3003 added handling of a signal interrupting the ibl hit path's jump, but the introduced
client.cleancallsig test could not test the ibl jump on AArchXX leading to the insertion of an
ASSERT_NOT_TESTED() for AARCHXX. Testing on a proprietary app with plain DR (debug build)
on AArch64 hit this case without problems. Thus, we can remove the ASSERT_NOT_TESTED()
for AARCH64.

The code inserted in PR #4797 was triggered and tested for AARCH64 on a proprietary app so
remove the ASSERT_NOT_TESTED() for AARCH64 from there as well.