i#4425: handle unspecified-by-the-app sigaction restorer for AArch64
Prevents a seg fault in the burst_aarch64_sys test that was caused by reading an unspecified
sigaction restorer in sig_has_restorer() in unix/signal.c. Does so by returning false early
in sig_has_restorer() for AArch64 when the SA_RESTORER flag is not set.

By preventing the seg fault, it also prevents the nested signal handling and consequently
the stack overflow in burst_aarch64_sys.

Issue: #4425