i#4316 a64 rseq: Port linux.rseq test to AArch64
Ports the linux.rseq test and its extensive inline assembly to
AArch64.  We use compiler-acquired C variable addresses prior to the
assembly sequences for simplicity.  This should be fine as there are
plenty of GPR's and DR handles inputs.

Tested natively and with forthcoming DR rseq mangling changes;
separated to keep diffs smaller.  These tests will be enabled in the
PR that adds DR handling.

Issue: #4316