i#4316 a64 rseq: Port rseq mangling to AArch64 (#4854)
Ports rseq mangling to AArch64.  This requires implementing
patch_mov_immed_arch() (done by leveraging existing raw encoding code
from exit stub support) and spilling extra scratch registers in
multiple places.

Expands translation support to cover the new mangling.  This includes
adding proper recognition of an mcontext base load, which was
incorrectly identified as an indirect branch target load on x86.

Ports the heuristic for finding the app's rseq TLS offset on attach
(when we did not see an rseq syscall) for AArch64 to look forward
instead of backward as on x86 where negative segment offsets are used.

Enables the rseq tests for AArch64.
Tested all 6 rseq tests manually on a machine with a newer kernel.
The Jenkins machine is too old and does not have rseq.
3 rseq tests do run on QEMU so those are enabled but until Jenkins is
upgraded we will not have perfect regression tests.

Updates the rseq docs to state that AArch64 is supported.
One final step is to support stores with writeback, which are seen in
real rseq sequences.  That will be done separately.  This completes
porting the x86 support to AArch64.

Issue: #4316