Add stricter free spill slot check.
This is needed when there are multiple instrumentation passes that
use drreg to spill reg values to slots. In this case, we need to
avoid spill slots that were already used in prior instrumentation
passes.

The added logic avoids any spill slot used by prior passes in the
current ilist. This is not completely optimal as a spill slot
may be available for use at some point in the ilist even if it is
already used elsewhere in the ilist. But we live with this
sub-optimality.

Issue: #3823