Fix check for prior pass usage of spill slot.
There was a bug in the check that led to a huge inefficiency in how spill slots were picked. It prevented any reuse of spill slots in a fragment, even if it wasn't being used anymore in the current instrumentation pass. This means that we'd run out of slots very quickly.

Instead of looking over all instrs in the ilist, a better check is to look at only the remaining ones (after 'where'). Any spill slot usage after 'where' must have been added by a prior instrumentation pass and not the current one.

Also, defer a get_scatter_gather_info call after it is determined that the instr is indeed a scatter or gather. This was causing some "instr_get_dst: ordinal invalid" issues in some sample clients.