Add pure asm test to verify store/load counts in expanded scatter/gather.
Fixes an issue with detecting kmovw, which leads to not enabling saving of zmm regs and opmask regs.

The store/load counts is currently broken due to a known drcachesim issue. Added a comment about that.

Also, the test had to use high numbered xmm registers to avoid conflict with xmm spill slot selection, which is very naive right now due to  missing support in drreg. Added a note for that as well.