i#3823 drreg: Avoid spill slot conflicts with earlier phases
Adds a stricter free spill slot check in drreg. This is needed when there are
multiple instrumentation passes that use drreg to spill reg values to slots.
In this case, we need to avoid spill slots that are in-use by prior instrumentation
passes. The added logic looks at the remaining instrs in the given ilist after
'where'. Any spill slot usage after 'where' must have been added by a prior
instrumentation pass and not the current one. To avoid spill slot conflicts,
we should avoid using such spill slots with pending usage. This is not completely
optimal as a spill slot may be available for use at some point in the ilist
even if it is used later in the ilist. But we live with this sub-optimality.

Adds new drreg_bb_properties_t (DRREG_HANDLE_MULTI_PHASE_SLOT_RESERVATIONS)
that enables the stricter free spill slot check. This is currently set by
drx_expand_scatter_gather.

Uses data area of label instrs marked with a special note to store information
about spill slot usage, which is used in the new stricter free spill slot check.

Adds documentation about the new option, highlighting that it's necessary to
enable this for multi-phase drreg reg reservations to work as intended, and
that it is not enabled by default due to performance cost.

Delays resetting bb_props to late in the instru2instru stage, in case stages
besides insertion need to read it.

Extends drreg-test to verify that spill slots used in later phases do not
conflict with those from earlier phases.

Fixes: #3823