Adjust slots reserved in drx.
The existing logic of updating slots reserved by drx if drx_expand_scatter_gather is invoked is racy. Instead, now we reserve 4 slots upfront in drx_init. This strategy overprovisions drx if no scatter gather instr is present, so we avoid it on Windows which is more sensitive to slot usage. On Linux, slots are more plentiful so this is okay. Also, we do not reserve the ideal number of slots (which is 6), to avoid underprovisioning other clients. This is okay as drx will fall back to using DR slots, which may be slower. All these effects are limited to scatter gather instrs that are split into their own bbs.