i#2985 scatter-gather: Expand scatter-gather instrs in tracing clients (#4863)
Invokes the existing drx_expand_scatter_gather to expand any scatter or gather instrs to scalar stores or loads, in app2app phase of drcachesim and other sample tracing clients.

Skips qword index/value variations of scatter/gather in 32-bit tests. Expansion for these variations is not yet supported in 32-bit. Added FIXME in tests for adding the missing pieces.

Defers a get_scatter_gather_info call after it is determined that the instr is indeed a scatter or gather. This was causing "instr_get_dst: ordinal invalid" issue in some sample clients.

Adds a pure asm test to verify exact store/load counts in expanded scatter/gather instrs. The store/load counts are currently incorrect due to a known drcachesim issue. Added a comment about that. Also, the test had to use high numbered xmm registers to avoid conflict with xmm spill slot selection, which is very naive right now due to missing spill support in drreg (i#3844). Added a note for that as well.

Also adds a drcachesim test for the existing drx-scattergather app.

Fixes an issue with detecting kmovw, which leads to not enabling saving of zmm and opmask regs on dispatch. Removes raw bytes check completely which didn't work well as it missed the kmovw case. As full_decode is set when clients process bbs, the decoded instr should be available anyway, and it is more straight-forward to just check the destinations directly.

Issue: #2985