i#4899: Fix drmemtrace drreg control flow discrepancies (#4901)
Adds recording of all app-valued registers at the skip point of
drmemtrace's instrumentation skips for filtered or global-maximum
offline trace options.  These registers are then restored prior to the
skip target, to ensure scratch register parity on all paths.

Tested on the precise basic block that triggered the bug when combined
with drreg limitations on which scratch registers to use.  Considered
trying to create a regression test out of this but there are many
complexities that make it non-trivial.

Fixes a pre-existing 32-bit ARM bug where a reg holding flags for the skip jump
was prematurely restored in the barrier.  Verified by manual testing.

Fixes #4899