i#2985 scatter-gather: Fix drcachesim and raw2trace issues.
Fixes issues in drcachesim and raw2trace because of which some
instructions and memrefs in the scatter/gather expansion were
not being instrumented. This led to lower than expected
basic_counts for data loads and stores.

In the scatter/gather expansion, all instrs have the same app
pc (this is unlike rep str expansion where the stringop has
the original app pc, and others have a fake one). So we need
to continue even if we see an instr with app pc same as the
last one we saw.

Scatter/gather expansion is different from rep str expansion
also because here we don't have any loop. So we need to make
sure to have a pc entry before each memref entry (raw2trace
requires this) by instrumenting each mov separately.

Another tricky behavior of scatter/gather expansion is that
the first non-label instruction in the expansion is a non-
app memref, which needs to be instrumented to output pc entry
(for scatter/gather completely suppressed due to  mask, we
still want a pc entry), but we don't want a memref entry for
that.

Disables instr bundling for scatter/gather expansions, so
that we can choose to not output a duplicate pc entry for
some instrs (specifically, the non-memrefs) in the online trace.

Adds drcachesim online test for allasm scatter gather app.

Adds an allasm app that uses repstr, and a drcachesim offline
and online test for this.

Issue: #2985