Skip address elision for all emulated instr cases.
This is done by identifying whether there is an emulation start/end marker in the ilist provided by instru_offline.
For the ilist provided by raw2trace, we check existence of scatter/gather, or rep stringop directly.

Also adds an assert to verify that first instr in expanded scatter/gather seq is non-app, so that we don't skip memrefs if it is app.

Also fixes format and adds issue number to TODO.