Track slot->reg instead of reg->slot data.
This is to handle the case where multiple phases spill the native app
value of some reg. In this case, the state restoration logic should
remember both slots, instead of just the latest one, because the
latest one may be recycled later.

Also adds test for this scenario.

Also adds test for the scenario when some app instr uses the stolen
reg x28 on AArch64 for some store, which looks like a spill and
corrupts the state restoration logic. Added a check that the
instr should be a meta-instr to be considered a spill/restore.